// src/Districtr/Districtr.types.ts
export interface AnyInputProps {
  /** The type of input for the option */
  type: 'rangeSlider' | 'colorPicker' | 'checkbox' | 'radio' | 'select'
  /** The name of the option */
  name: string
  /** The property that the input controls on the tool */
  property: string
  /** The configuration for the input */
  config: ColorPickerProps | RangeSliderProps
}

export interface AnyToolProps {
  /** The name of the tool */
  name: string
  /** The icon for the tool */
  icon: string
  /** The tooltip for the tool */
  tooltip: string
  /** The cursor for the tool */
  cursor: string
  /** The keyboard shortcut for the tool. */
  shortcut?: string
  /** Is the tool enabled
   * @default true
   */
  enabled?: boolean
}

export interface BrushToolProps extends AnyToolProps {
  /** The size of the brush in meters.
   * @default 5000
   */
  size: number
  /** The brush tool options configuration */
  options?: {
    /** the inputs and their configuration */
    inputs: AnyInputProps[]
  }
}

export interface ToolsConfigProps {
  /** The default tool */
  brush?: BrushToolProps
  eraser?: BrushToolProps
  pan?: AnyToolProps
}